# docs/openapi.yaml
openapi: 3.0.2
info:
  title: ''
  version: ''
paths:
...
  /api/user/login/:
    post:
      operationId: createLogin
      description: 'login'
      parameters: []
      requestBody:
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/login'
          application/x-www-form-urlencoded:
            schema: {}
          multipart/form-data:
            schema: {}
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: ''
      tags:
      - api
...
  /api/book/reserve-book/:
    post:
      operationId: createReservation
      description: 'رزرو کردن کتاب'
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Reservation'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Reservation'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
          description: ''
      tags:
      - api
...
components:
  schemas:
...
    Reservation:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        user:
          type: string
          readOnly: true
        book:
          type: string
        reservation_date:
          type: string
          format: date-time
          readOnly: true
        return_date:
          type: string
          format: date-time
          nullable: true
        rid:
          type: string
          readOnly: true
      required:
      - book
...
    login:
      type: object
      properties:
        password:
          type: string
          writeOnly: true
        phone_number:
          type: string
          maxLength: 11
      required:
      - password
      - phone_number
...